Move EndTimeContext to its own file
Noticed in #848 that it was creating a dependency cycle.
