+++ packages/core/src/components/EndTimeContext.tsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const EndTimeContext = createContext<number>(Infinity);
