+++ packages/core/src/components/FlightResponseTabNetwork.tsx
@@ -1,7 +1,7 @@
 import { Chunk, FlightResponse, isReference } from "../react/ReactFlightClient";
 import React, { memo, useContext, useEffect, useRef, useState } from "react";
 import * as d3 from "d3";
-import { EndTimeContext } from "./ViewerStreams";
+import { EndTimeContext } from "./EndTimeContext";
 
 interface Node extends d3.SimulationNodeDatum {
   chunk: Chunk;
